src/requirements.txt
awsiotsdk==1.21.0
boto3==1.34.144
python-dotenv==1.0.0